Move isc types to sui_go
