--- packages/isc/sui/mock_contract/Move.lock
@@ -1,37 +1,0 @@
-# @generated by Move, please check-in and do not edit manually.
-
-[move]
-version = 1
-manifest_digest = "E5EA3EEAAF9D683B2BB3568861A10871B4AFBB37EB2DF36D3DD272C656A496B5"
-deps_digest = "060AD7E57DFB13104F21BE5F5C3759D03F0553FC3229247D9A7A6B45F50D03A3"
-dependencies = [
-  { name = "MoveStdlib" },
-  { name = "Stardust" },
-  { name = "Sui" },
-]
-
-[[move.package]]
-name = "MoveStdlib"
-source = { git = "git@github.com:iotaledger/iota.git", rev = "l1sc/isc-models", subdir = "crates/sui-framework/packages/move-stdlib" }
-
-[[move.package]]
-name = "Stardust"
-source = { git = "git@github.com:iotaledger/iota.git", rev = "l1sc/isc-models", subdir = "crates/sui-framework/packages/stardust" }
-
-dependencies = [
-  { name = "MoveStdlib" },
-  { name = "Sui" },
-]
-
-[[move.package]]
-name = "Sui"
-source = { git = "git@github.com:iotaledger/iota.git", rev = "l1sc/isc-models", subdir = "crates/sui-framework/packages/sui-framework" }
-
-dependencies = [
-  { name = "MoveStdlib" },
-]
-
-[move.toolchain-version]
-compiler-version = "1.22.0"
-edition = "legacy"
-flavor = "sui"
