+++ sui-go/iscmove/types/mock_contract/bytecode.json
