+++ sui-go/iscmove/types/mock_contract/embed.go
